Group management implementation.
Alright, this is the big last major bit for group chat support.  You can
now create new groups, invite people, and when they leave the group
membership is reflected properly.

This all still seems a bit brittle and I can't figure out if it's issues
in the code or in signald, but in general it seems to work!