Fixed an out-of-date comment.
Times like this I sympathize with those who disagree with heavily
documenting code.

The old comment reflects when I was identifying chats by the order in
which we discovered them, but I've since moved to hashing the group ID
itself to get a 32-bit number that's probably unique enough for our
needs.