Refactoring round two - the breakup.
Alright, this is the big one.  Here we split up libsignald.c into a
bunch of modules with some reasonably cohesive concerns, and the little
internal bits buried.  I won't claim it's perfect, but I think it's
definitely an improvement.

This is definitely the less risky of the two refactorings.  There's a
couple of slightly riskier changes here and there where I had to move
some code around, but mostly this is a simple breakup of the code along
functional lines.  It's the preceding commit that really bears some
scrutiny...