Fix behaviour with persistent chats.
When Pidgin sets a chat as persistent, we don't get leave messages when
the chat is closed, but we *do* get join messages!

So we need to look for the chat (right now we have to look for it by
name, so no multiple chats with the same name), and if we know about it,
skip creating a new group.

This allows the user to "Conversation -> Add..." for joined chats, and
then mark the chat persistent so that they don't leave it if they
close it.