Switch to conv_chat_write for self-messages.
When sending a message to a group chat, we have to manually echo the
message back into the chat because Signal doesn't send it back to us.

Previously we used purple_serv_got_chat_in, which is incorrect.  This
change fixes the method to use purple_conv_chat_write, which is the
right way to perform this function.