External attachment implementation.
This changeset adds a few new settings that allow libpurple-signald to
work in environments where attachments must be accessed via an external
web server or other mechanism.

In particular, it adds three new settings:

* external-attachments (boolean)
* external-attachments-dir (string)
* external-attachments-url (string)

If `external-attachments` is set to false, then libpurple-signald
behaves as it did before, inlining the images directly.

If it's set to true, then the library copies the attachment to the
directory indicated by `external-attachments-dir` with an appropriate
file extension (detected by libmagic).

It then emits into the conversation a URL constructed by appending the
basename and extension to the value specified in
`external-attachments-url`.

The idea, here, is that a user can set up a web server to host
attachments, and then the library will place attachments there and
provide appropriate links.  This is important for text-mode clients like
bitlbee that don't support direct image inlining.