Support both old and new signald protocols
There's a few major pieces to this commit.

First, there's the protocol version detection, and some associated
refactoring.

In general, the initialization cycle for the client is now wound up in
the message send/receive loop.  So once a 'subscribed' event is
received, a series of chained requests occur to:

1. Check the protocol version (I'll get to this in a second)
2. When the version check returns, request the contact list
2. When the contact list comes back, request the group list
3. When the group list comes back, indicate we're initialized

So, how do we check the protocol version?  Well... it's kinda ugly.

The `get_user` call is used to get information about a Signal user.  In
the old version of the protocol, it took a `recipientNumber`.  In the
new version, it takes a `recipientAddress`.

So, we invoke `get_user` to get information about the current user, and
in the call, we provide a `recipientAddress`.

If the call succeeds and we get a `user` or `user_not_registered`
message type back, it's the new protocol version.

If the call returns an `unexpected_error`, and we're still initializing
and the error was from a `get_user` call, we assume this is an old
version of signald.

We track whether we're dealing with a legacy signald instance with a
flag in the SignaldAccount.

Finally, throughout the codebase, where necessary, the `legacy_protocol`
flag is checked before parsing or populating various JSON payloads.

Phew!  There, that's it.  And it even seems to work.