Refactoring before some behavioural changes.
Ultimately my goal is to improve group chat auto-join behaviour by
making things a bit more configurable.  In particular, I want the plugin
to support having a record of an existing conversation but it not being
joined yet, so that joins can optionally be done on-demand from the
buddy list.

But first I needed to refactor a bunch of this code (again).

Prior to this commit, information about the group chats--and in
particular the user list for those chats--was being carried around by
Pidgin PurpleConversation instances.  This meant that a
PurpleConversation instance had to be created when we discovered a group
chat.

Obviously that doesn't work if there isn't a 1-to-1 relationship between
PurpleConversation instances and Signal groups.

So this change adds a bunch of additional bookkeeping for groups so that
we track this metadata separately from the Pidgin internals.

The next step will then be to implement the behaviour enhancements.  But
since this rewiring is so significant, I wanted to get it done, first,
before pushing on, just in case I really broke things...