Profound refactoring of message handling.
As the initialization logic threatens to get more complicated, this
rewrite transforms the message handling loop into a formal state machine
with named states and transitions.

This makes it a lot easier to express the initialization cycle, among
other things, and also naturally expresses the primary event handling
loop as well.