Handle link errors in state machine
This enhances the state machine a bit so that a given state may have
multiple transitions for a given message type, and then a filter can be
registered with the transition to determine if that transition should be
taken.

This mechanism is then used to identify link errors at startup (which
then triggers a re-linking attempt).

This now means the link error detection is incorporated into the state
machine.