Control whether chats are opened at startup.
This addresses issue #13 by adding a new setting (auto-join-group-chats)
that controls whether or not group chats are automatically joined/opened
at startup.

If the setting is turned off, group chats are only opened if a message
is received.  I've opted to do this to ensure no messages are lost (you
need the conversation open so you have a place to put the messages as
they arrive!)

If the setting is turned on, group chats are opened at startup (the
previous behaviour), which is preferable for bitlbee.

The default is for this setting to be turned *off*.