Basic quoted message support.
Take out the quoted text, split it by newlines, prepend each line with
"> " to indicate it's a quote, then emit it, followed by a newline and
the reply text.

This looks good in Bitlbee and just okay in Pidgin.  I'm starting to
wonder if multi-line messages should be split and emitted in Pidgin,
generally, but for now this is in keeping with how these types of
messages are formatted today.