Remove legacy protocol support
Both signald and its clients should be updated regularly to stay compatible
with Signal itself, so there is no need to support legacy protocols for
more than a short transition period.

Also, in this case, it seems libpurple-signald can be confused into using
the legacy protocol even with a recent signald, causing the communcation to
fail and losing messages.